# [dnm] tracing: investigate perf regression under sql.txn_stats.sample_rate=1