--- a/out.1
+++ b/out.1
@@ -1,0 +1,105 @@
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    315279 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    307033 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    324020 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    313610 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    303585 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    318193 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    313108 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    322368 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    307665 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    325641 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    307798 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    291304 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    312540 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    319267 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    323648 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    314294 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    331358 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    299547 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    318578 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    298624 ns/op
+PASS
+goos: linux
+goarch: amd64
+pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    320333 ns/op
+PASS