--- a/out.1
+++ b/out.1
@@ -1,105 +1,105 @@
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    315279 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    316450 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    307033 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    331154 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    324020 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    291603 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    313610 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    300284 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    303585 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    311225 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    318193 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    317917 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    313108 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    315304 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    322368 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    318753 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    307665 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    285229 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    325641 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    304026 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    307798 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    294862 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    291304 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    295566 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    312540 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    327735 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    319267 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    317495 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    323648 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    290513 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    314294 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    301583 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    331358 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    307657 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    299547 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    289831 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    318578 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    298596 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    298624 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    313526 ns/op	         8.00 spans/op
 PASS
 goos: linux
 goarch: amd64
 pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
-BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    320333 ns/op
+BenchmarkKV/Scan/SQL/rows=1/sample_rate=X-16         	    5000	    320642 ns/op	         8.00 spans/op
 PASS