… pkg/ccl/streamingccl/streamclient/BUILD.bazel
… pkg/ccl/streamingccl/streamclient/client_test.go
… pkg/ccl/streamingccl/streamclient/cockroach_sinkless_replication_client_test.go
… pkg/ccl/streamingccl/streamclient/main_test.go
… pkg/ccl/streamingccl/streamclient/stream_client.go
… pkg/ccl/streamingccl/streamingest/BUILD.bazel
… pkg/ccl/streamingccl/streamingest/stream_ingestion_frontier_processor_test.go
… pkg/ccl/streamingccl/streamingest/stream_ingestion_job_test.go
… pkg/ccl/streamingccl/streamingest/stream_ingestion_planning.go
… pkg/ccl/streamingccl/streamingest/stream_ingestion_processor.go
… pkg/ccl/streamingccl/streamingest/stream_ingestion_processor_test.go
… pkg/ccl/streamingccl/streamingest/stream_ingestion_test.go
… pkg/ccl/streamingccl/streamingtest/BUILD.bazel
… pkg/ccl/streamingccl/streamingtest/encoding.go
… pkg/ccl/streamingccl/streamingtest/replication_helpers.go
… pkg/ccl/streamingccl/streamingutils/utils_test.go
… pkg/ccl/streamingccl/streamproducer/BUILD.bazel
… pkg/ccl/streamingccl/streamproducer/replication_stream_test.go
… pkg/kv/kvserver/client_merge_test.go
… pkg/kv/kvserver/closedts/sidetransport/sender.go
… pkg/kv/kvserver/replica.go
… pkg/kv/kvserver/replica_command.go
… pkg/roachpb/batch.go
… pkg/server/server.go
… pkg/server/status/runtime.go
… pkg/sql/colexec/BUILD.bazel
… pkg/sql/colexec/colbuilder/execplan.go
… pkg/sql/colexec/colbuilder/execplan_test.go
… pkg/sql/colexec/columnarizer.go
… pkg/sql/colexec/materializer.go