Old is `sample_rate=0`, new is `sample_rate=1`.

```
$ ./cmd.sh
name                                 old time/op  new time/op  delta
KV/Scan/SQL/rows=1/sample_rate=X-16   282µs ± 7%   336µs ± 9%  +18.97%  (p=0.000 n=10+10)
```

Touches https://github.com/cockroachdb/cockroach/issues/59379.