@tbg 2021-03-04
Updated the PR on top of your recent work, @yuzefovich 

@irfansharif as we discussed yesterday, mind taking a look here with me? The branch should be relatively clean now. I added a microbenchmark that tries to capture exactly the overhead that would be incurred from setting the sampling rate from zero to one. It incurs a cost, but I don't see anything that can explain the slow down I see via `./cmp.sh`.

@yuzefovich could you scrutinize the microbenchmark? Is there any significant work that is done as a result of propagating component stats that it doesn't capture?

```
$ go test -run - -bench SpanStats -count 1 ./pkg/util/tracing
goos: linux
goarch: amd64
pkg: github.com/cockroachdb/cockroach/pkg/util/tracing
BenchmarkSpanStats/autoCollection=false,stats=0-16         	 280952	     4326 ns/op
BenchmarkSpanStats/autoCollection=false,stats=1-16         	 158040	     7901 ns/op

BenchmarkSpanStats/autoCollection=true,stats=0-16          	 471117	     2631 ns/op
BenchmarkSpanStats/autoCollection=true,stats=1-16          	 242348	     5919 ns/op
PASS
ok  	github.com/cockroachdb/cockroach/pkg/util/tracing	8.414s
```

The case `autoCollection=false,stats=1` should roughly reflect the SQL
usage in which stats are created, transported via a recording, ingested
into a parent, and then retrieved and unmarshaled. The case `false,stats=0`
should correspond to the non-sampled case. We are thus expecting roughly
an overhead of 4.3µs/op when not sampled, 8µs/op when sampled, i.e.
changing the sampling rate from 0 to 1 should add ~4µs/op.
However, the macro-level benchmark shows much more:

```
KV/Scan/SQL/rows=1/sample_rate=X-16   268µs ± 9%   314µs ± 7%  +17.03%  (p=0.000 n=21+21)
```

That's a drop of 46µs incurred by changing the sampling rate from
zero to one, much past what can be explained by this microbenchmark.

The autoCollection=true numbers avoid an extra unmarshal-remarshal cycle
of the recording. This is a possible optimization for SQL to investigate
in the case in which the flow lives wholly on the gateway.
Unfortunately, it is finicky to get right and causes lots of special
casing, as SQL currently relies on all trace data arriving at the
DistSQL RowReceiver via TrailingMeta (i.e. marshalled recordings).