@tbg 2021-03-04
This is also consistent with this:

With `RecordStructured` turned into a noop:

```
name                                 old time/op  new time/op  delta
KV/Scan/SQL/rows=1/sample_rate=X-16   261µs ±10%   292µs ±10%  +12.08%  (p=0.000 n=20+21)
name                                 old time/op  new time/op  delta
KV/Scan/SQL/rows=1/sample_rate=X-16   267µs ±11%   293µs ± 6%  +9.40%  (p=0.000 n=21+21)
```

With `RecordStructured` doing its job:

```
name                                 old time/op  new time/op  delta
KV/Scan/SQL/rows=1/sample_rate=X-16   271µs ± 9%   310µs ± 8%  +14.69%  (p=0.000 n=21+21)
name                                 old time/op  new time/op  delta
KV/Scan/SQL/rows=1/sample_rate=X-16   276µs ± 7%   316µs ± 8%  +14.48%  (p=0.000 n=21+20)
```

If throwing away the `ComponentStats` offered to the Span, they won't get
marshalled or unmarshalled or even put on the wire. And yet, there's a 10%
performance hit for the setting. What is the expensive thing we're doing?
Creating more spans?