@tbg 2021-03-04
@yuzefovich  I added `span/op` counts to the micro and SELECT benchmark I'm running. This shows that in `./cmp.sh` benchmark, we're creating 8 spans when sampling, 7 when not sampling.

That one additional trace span is unlikely to cost ~40us, but avoiding
it would establish more parity between the tests. Can we do that, at least as a hack to see what the effect would be?