@tbg 2021-03-04
I tracked down the extra span, and it's this one:

https://github.com/cockroachdb/cockroach/blob/a9b582fcc94fbc022a2deb5b8ccda885171cc189/pkg/sql/colfetcher/colbatch_scan.go#L92-L96

I *optimized it out* :sunglasses: but as expected it didn't make it or break it:

```diff
diff --git a/pkg/sql/colfetcher/colbatch_scan.go b/pkg/sql/colfetcher/colbatch_scan.go
index cae59582b8..6a69ff7f96 100644
--- a/pkg/sql/colfetcher/colbatch_scan.go
+++ b/pkg/sql/colfetcher/colbatch_scan.go
@@ -92,7 +92,12 @@ func (s *ColBatchScan) Next(ctx context.Context) coldata.Batch {
                if execinfra.ShouldCollectStats(s.ctx, s.flowCtx) {
                        // We need to start a child span so that the only contention events
                        // present in the recording would be because of this cFetcher.
-                       s.ctx, s.tracingSpan = execinfra.ProcessorSpan(s.ctx, "colbatchscan")
+                       sp := tracing.SpanFromContext(s.ctx)
+                       if sp == nil {
+                               panic("oops")
+                       }
+                       s.tracingSpan = sp
+                       // s.ctx, s.tracingSpan = execinfra.ProcessorSpan(s.ctx, "colbatchscan")
                }
        }
        bat, err := s.rf.NextBatch(s.ctx)
```

```
name                                 old time/op   new time/op   delta
KV/Scan/SQL/rows=1/sample_rate=X-16    266µs ± 7%    316µs ± 7%  +18.68%  (p=0.000 n=21+21)

name                                 old spans/op  new spans/op  delta
KV/Scan/SQL/rows=1/sample_rate=X-16     7.00 ± 0%     7.00 ± 0%     ~     (all equal)
```