@tbg 2021-03-04
So after all of this, it seems to me that there's evidence that something in SQL is slower when stats collection is enabled, *even when tracing just drops all payloads on the floor and is thus definitely not incurring overhead for them*, and *even when the number of trace spans created is identical*.