@yuzefovich 2021-03-05
Thanks Tobi. I think the benchmark you wrote is a reasonable representation of how SQL propagates stats.

> And yet, there's a 10% performance hit for the setting. What is the expensive thing we're doing? Creating more spans?

These are the questions I would like to know the answer to as well :)

As you found, we are creating an extra span in `colbatch_scan`, and I don't think we can get around that (without it we could be confusing contention events from different table readers running on the same node I think, to be confirmed), yet it doesn't appear to be the culprit.

I'll keep on looking at things to optimize in the SQL land (already found one #61529 which is responsible for a third of so of the performance hit). What puzzles me a bit is that usually I'm using profiles to guide the code optimization, but this time I'm not able to do that. I tried increasing CPU sampling rate (from default 100hz to 40000hz), and still the profiles aren't very helpful. The only strategy that appears to be working is simply commenting out parts that run once sampling is enabled and seeing whether it has any impact on the microbenchmark.